refactor: Move GetConst into ConstantFolder
